Added Dail-Madsen simulator with diffusive movement
Added the simulation script ("Simulation_model.R"), which sources
functions for calculating Dail-Madsen (DM) or CJS data for each
individual site.  I've only implemented the DM simulator so far.  I'd
like confirm (1) that this coding set-up will be ok and readable, (2)
that the model with diffusive movement is ok, (3) that the input
parameters cover the range of things we'll want to explore.  When I get
confirmation, I'll then do the CJS simulator.